consolidate mangle directory
