--- pydl/pydlutils/mangle/__init__.py
@@ -1,8 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-"""
-This subpackage corresponds to the mangle directory of idlutils.
-"""
-from .is_cap_used import is_cap_used
-from .read_fits_polygons import read_fits_polygons
-from .set_use_caps import set_use_caps
