--- pydl/pydlutils/mangle/is_cap_used.py
@@ -1,17 +1,0 @@
-# Licensed under a 3-clause BSD style license - see LICENSE.rst
-# -*- coding: utf-8 -*-
-def is_cap_used(use_caps,i):
-    """Returns ``True`` if a cap is used.
-
-    Parameters
-    ----------
-    use_caps : :class:`int`
-        Bit mask indicating which cap is used.
-    i : :class:`int`
-        Number indicating which cap we are interested in.
-
-    Returns
-    -------
-    is_cap_used : :class:`bool`
-    """
-    return (use_caps & 1 << i) != 0
